feat: update singleton scoped resolution
Update resolver builders to not to use JIT on singleton scoped flows
