+++ packages/container/libraries/core/src/planning/calculations/buildFourConstructorArgumentResolver.ts
@@ -10,23 +10,23 @@ import {
 import { type ConstructorNoParamNode } from '../models/ConstructorNoParamNode.js';
 import { type InstanceBindingNode } from '../models/InstanceBindingNode.js';
 import { type PlanServiceNode } from '../models/PlanServiceNode.js';
-import { resolveTwo } from './resolveTwo.js';
+import { resolveFour } from './resolveFour.js';
 
 /**
- * Same rationale as buildZeroConstructorArgumentsResolverOnCsp, but for
- * two-argument instance bindings. Equivalent to
- * `buildConstructorArgumentsResolver` with `resolveTwo`, but implemented with a
- * plain closure instead of the `Function` constructor, so it works in
+ * Same rationale as buildZeroConstructorArgumentsResolver, but for
+ * four-argument instance bindings. Equivalent to
+ * `buildConstructorArgumentsResolverJit` with `resolveFour`, but implemented
+ * with a plain closure instead of the `Function` constructor, so it works in
  * environments enforcing a strict Content Security Policy (no
  * `unsafe-eval`).
  *
  * When the bound class has no properties to inject,
  * `node.classMetadata.properties` is empty and the returned `resolveNode`
  * never performs any property related check, matching the zero-property
- * fast path performance of `buildConstructorArgumentsResolver` with
- * `resolveTwo`.
+ * fast path performance of `buildConstructorArgumentsResolverJit` with
+ * `resolveFour`.
  */
-export function buildTwoConstructorArgumentResolverOnCsp<TActivated>(
+export function buildFourConstructorArgumentResolver<TActivated>(
   node: InstanceBindingNode<TActivated, InstanceBinding<TActivated>>,
   resolveActivations?: (
     params: ResolutionParams,
@@ -44,12 +44,30 @@ export function buildTwoConstructorArgumentResolverOnCsp<TActivated>(
         const resolvedValue1: unknown = (
           node.constructorParams[1] as PlanServiceNode | ConstructorNoParamNode
         ).resolve(params);
+        const resolvedValue2: unknown = (
+          node.constructorParams[2] as PlanServiceNode | ConstructorNoParamNode
+        ).resolve(params);
+        const resolvedValue3: unknown = (
+          node.constructorParams[3] as PlanServiceNode | ConstructorNoParamNode
+        ).resolve(params);
 
-        return resolveTwo(
+        return resolveFour(
           resolvedValue0,
           resolvedValue1,
-          (resolvedValue0: unknown, resolvedValue1: unknown): TActivated =>
-            new node.binding.implementationType(resolvedValue0, resolvedValue1),
+          resolvedValue2,
+          resolvedValue3,
+          (
+            resolvedValue0: unknown,
+            resolvedValue1: unknown,
+            resolvedValue2: unknown,
+            resolvedValue3: unknown,
+          ): TActivated =>
+            new node.binding.implementationType(
+              resolvedValue0,
+              resolvedValue1,
+              resolvedValue2,
+              resolvedValue3,
+            ),
         );
       };
     }
@@ -63,17 +81,32 @@ export function buildTwoConstructorArgumentResolverOnCsp<TActivated>(
       const resolvedValue1: unknown = (
         node.constructorParams[1] as PlanServiceNode | ConstructorNoParamNode
       ).resolve(params);
+      const resolvedValue2: unknown = (
+        node.constructorParams[2] as PlanServiceNode | ConstructorNoParamNode
+      ).resolve(params);
+      const resolvedValue3: unknown = (
+        node.constructorParams[3] as PlanServiceNode | ConstructorNoParamNode
+      ).resolve(params);
 
-      return resolveTwo(
+      return resolveFour(
         resolvedValue0,
         resolvedValue1,
+        resolvedValue2,
+        resolvedValue3,
         (
           resolvedValue0: unknown,
           resolvedValue1: unknown,
+          resolvedValue2: unknown,
+          resolvedValue3: unknown,
         ): Resolved<TActivated> =>
           resolveActivations(
             params,
-            new node.binding.implementationType(resolvedValue0, resolvedValue1),
+            new node.binding.implementationType(
+              resolvedValue0,
+              resolvedValue1,
+              resolvedValue2,
+              resolvedValue3,
+            ),
           ),
       );
     };
@@ -106,19 +139,31 @@ export function buildTwoConstructorArgumentResolverOnCsp<TActivated>(
       const resolvedValue1: unknown = (
         node.constructorParams[1] as PlanServiceNode | ConstructorNoParamNode
       ).resolve(params);
+      const resolvedValue2: unknown = (
+        node.constructorParams[2] as PlanServiceNode | ConstructorNoParamNode
+      ).resolve(params);
+      const resolvedValue3: unknown = (
+        node.constructorParams[3] as PlanServiceNode | ConstructorNoParamNode
+      ).resolve(params);
 
-      return resolveTwo(
+      return resolveFour(
         resolvedValue0,
         resolvedValue1,
+        resolvedValue2,
+        resolvedValue3,
         (
           resolvedValue0: unknown,
           resolvedValue1: unknown,
+          resolvedValue2: unknown,
+          resolvedValue3: unknown,
         ): Resolved<TActivated> =>
           finalizeInstance(
             params,
             new node.binding.implementationType(
               resolvedValue0,
               resolvedValue1,
+              resolvedValue2,
+              resolvedValue3,
             ) as SyncResolved<TActivated> & Record<string | symbol, unknown>,
           ),
       );
@@ -151,19 +196,31 @@ export function buildTwoConstructorArgumentResolverOnCsp<TActivated>(
     const resolvedValue1: unknown = (
       node.constructorParams[1] as PlanServiceNode | ConstructorNoParamNode
     ).resolve(params);
+    const resolvedValue2: unknown = (
+      node.constructorParams[2] as PlanServiceNode | ConstructorNoParamNode
+    ).resolve(params);
+    const resolvedValue3: unknown = (
+      node.constructorParams[3] as PlanServiceNode | ConstructorNoParamNode
+    ).resolve(params);
 
-    return resolveTwo(
+    return resolveFour(
       resolvedValue0,
       resolvedValue1,
+      resolvedValue2,
+      resolvedValue3,
       (
         resolvedValue0: unknown,
         resolvedValue1: unknown,
+        resolvedValue2: unknown,
+        resolvedValue3: unknown,
       ): Resolved<TActivated> =>
         finalizeInstance(
           params,
           new node.binding.implementationType(
             resolvedValue0,
             resolvedValue1,
+            resolvedValue2,
+            resolvedValue3,
           ) as SyncResolved<TActivated> & Record<string | symbol, unknown>,
         ),
     );
