+++ packages/container/libraries/core/src/planning/calculations/buildThreeResolvedValueArgumentResolver.spec.ts
@@ -5,7 +5,7 @@ import { type ResolutionParams } from '../../resolution/models/ResolutionParams.
 import { type Resolved } from '../../resolution/models/Resolved.js';
 import { type PlanServiceNode } from '../models/PlanServiceNode.js';
 import { type ResolvedValueBindingNode } from '../models/ResolvedValueBindingNode.js';
-import { buildTwoResolvedValueArgumentResolverOnCsp } from './buildTwoResolvedValueArgumentResolverOnCsp.js';
+import { buildThreeResolvedValueArgumentResolver } from './buildThreeResolvedValueArgumentResolver.js';
 
 class TestFixtures {
   public static get node(): ResolvedValueBindingNode<
@@ -13,10 +13,10 @@ class TestFixtures {
   > {
     return {
       binding: {
-        factory: (value0: unknown, value1: unknown): string =>
-          `factory:${String(value0)}:${String(value1)}`,
+        factory: (value0: unknown, value1: unknown, value2: unknown): string =>
+          `factory:${String(value0)}:${String(value1)}:${String(value2)}`,
         metadata: {
-          arguments: [Symbol(), Symbol()],
+          arguments: [Symbol(), Symbol(), Symbol()],
         },
       },
       params: [],
@@ -31,6 +31,7 @@ class TestFixtures {
 function buildParamsFixture(
   resolve0: () => unknown,
   resolve1: () => unknown,
+  resolve2: () => unknown,
 ): PlanServiceNode[] {
   return [
     {
@@ -39,10 +40,13 @@ function buildParamsFixture(
     {
       resolve: resolve1,
     } as Partial<PlanServiceNode> as PlanServiceNode,
+    {
+      resolve: resolve2,
+    } as Partial<PlanServiceNode> as PlanServiceNode,
   ];
 }
 
-describe(buildTwoResolvedValueArgumentResolverOnCsp, () => {
+describe(buildThreeResolvedValueArgumentResolver, () => {
   describe('when called, and resolveActivations is not provided', () => {
     describe('when called, and node.params is populated after the resolver is built', () => {
       let result: unknown;
@@ -52,12 +56,13 @@ describe(buildTwoResolvedValueArgumentResolverOnCsp, () => {
           ResolvedValueBinding<string>
         > = TestFixtures.node;
         const resolveNode: (params: ResolutionParams) => Resolved<string> =
-          buildTwoResolvedValueArgumentResolverOnCsp(nodeFixture);
+          buildThreeResolvedValueArgumentResolver(nodeFixture);
 
         nodeFixture.params.push(
           ...buildParamsFixture(
             (): string => 'value-0',
             (): string => 'value-1',
+            (): string => 'value-2',
           ),
         );
 
@@ -65,7 +70,7 @@ describe(buildTwoResolvedValueArgumentResolverOnCsp, () => {
       });
 
       it('should call the factory with the resolved arguments', () => {
-        expect(result).toBe('factory:value-0:value-1');
+        expect(result).toBe('factory:value-0:value-1:value-2');
       });
     });
   });
@@ -79,7 +84,7 @@ describe(buildTwoResolvedValueArgumentResolverOnCsp, () => {
           ResolvedValueBinding<string>
         > = TestFixtures.node;
         const resolveNode: (params: ResolutionParams) => Resolved<string> =
-          buildTwoResolvedValueArgumentResolverOnCsp(
+          buildThreeResolvedValueArgumentResolver(
             nodeFixture,
             (
               _params: ResolutionParams,
@@ -91,6 +96,7 @@ describe(buildTwoResolvedValueArgumentResolverOnCsp, () => {
           ...buildParamsFixture(
             (): string => 'value-0',
             (): string => 'value-1',
+            (): string => 'value-2',
           ),
         );
 
@@ -98,29 +104,43 @@ describe(buildTwoResolvedValueArgumentResolverOnCsp, () => {
       });
 
       it('should call the factory with the resolved arguments', () => {
-        expect(result).toBe('factory:value-0:value-1');
+        expect(result).toBe('factory:value-0:value-1:value-2');
       });
     });
 
-    describe.each<[string, () => unknown, () => unknown]>([
+    describe.each<[string, () => unknown, () => unknown, () => unknown]>([
       [
-        'both arguments resolve asynchronously',
+        'all arguments resolve asynchronously',
         async (): Promise<string> => 'value-0',
         async (): Promise<string> => 'value-1',
+        async (): Promise<string> => 'value-2',
       ],
       [
         'the first argument resolves asynchronously',
         async (): Promise<string> => 'value-0',
         (): string => 'value-1',
+        (): string => 'value-2',
       ],
       [
         'the second argument resolves asynchronously',
         (): string => 'value-0',
         async (): Promise<string> => 'value-1',
+        (): string => 'value-2',
+      ],
+      [
+        'the third argument resolves asynchronously',
+        (): string => 'value-0',
+        (): string => 'value-1',
+        async (): Promise<string> => 'value-2',
       ],
     ])(
       'when called, and %s',
-      (_: string, resolve0: () => unknown, resolve1: () => unknown) => {
+      (
+        _: string,
+        resolve0: () => unknown,
+        resolve1: () => unknown,
+        resolve2: () => unknown,
+      ) => {
         let result: unknown;
 
         beforeAll(async () => {
@@ -128,7 +148,7 @@ describe(buildTwoResolvedValueArgumentResolverOnCsp, () => {
             ResolvedValueBinding<string>
           > = TestFixtures.node;
           const resolveNode: (params: ResolutionParams) => Resolved<string> =
-            buildTwoResolvedValueArgumentResolverOnCsp(
+            buildThreeResolvedValueArgumentResolver(
               nodeFixture,
               (
                 _params: ResolutionParams,
@@ -136,13 +156,15 @@ describe(buildTwoResolvedValueArgumentResolverOnCsp, () => {
               ): Resolved<string> => resolvedValue,
             );
 
-          nodeFixture.params.push(...buildParamsFixture(resolve0, resolve1));
+          nodeFixture.params.push(
+            ...buildParamsFixture(resolve0, resolve1, resolve2),
+          );
 
           result = await resolveNode(TestFixtures.params);
         });
 
         it('should call the factory with the resolved arguments', () => {
-          expect(result).toBe('factory:value-0:value-1');
+          expect(result).toBe('factory:value-0:value-1:value-2');
         });
       },
     );
