+++ packages/container/libraries/core/src/planning/calculations/buildTwoConstructorArgumentResolver.ts
@@ -10,23 +10,23 @@ import {
 import { type ConstructorNoParamNode } from '../models/ConstructorNoParamNode.js';
 import { type InstanceBindingNode } from '../models/InstanceBindingNode.js';
 import { type PlanServiceNode } from '../models/PlanServiceNode.js';
-import { resolveFour } from './resolveFour.js';
+import { resolveTwo } from './resolveTwo.js';
 
 /**
- * Same rationale as buildZeroConstructorArgumentsResolverOnCsp, but for
- * four-argument instance bindings. Equivalent to
- * `buildConstructorArgumentsResolver` with `resolveFour`, but implemented
- * with a plain closure instead of the `Function` constructor, so it works in
+ * Same rationale as buildZeroConstructorArgumentsResolver, but for
+ * two-argument instance bindings. Equivalent to
+ * `buildConstructorArgumentsResolverJit` with `resolveTwo`, but implemented with a
+ * plain closure instead of the `Function` constructor, so it works in
  * environments enforcing a strict Content Security Policy (no
  * `unsafe-eval`).
  *
  * When the bound class has no properties to inject,
  * `node.classMetadata.properties` is empty and the returned `resolveNode`
  * never performs any property related check, matching the zero-property
- * fast path performance of `buildConstructorArgumentsResolver` with
- * `resolveFour`.
+ * fast path performance of `buildConstructorArgumentsResolverJit` with
+ * `resolveTwo`.
  */
-export function buildFourConstructorArgumentResolverOnCsp<TActivated>(
+export function buildTwoConstructorArgumentResolver<TActivated>(
   node: InstanceBindingNode<TActivated, InstanceBinding<TActivated>>,
   resolveActivations?: (
     params: ResolutionParams,
@@ -44,30 +44,12 @@ export function buildFourConstructorArgumentResolverOnCsp<TActivated>(
         const resolvedValue1: unknown = (
           node.constructorParams[1] as PlanServiceNode | ConstructorNoParamNode
         ).resolve(params);
-        const resolvedValue2: unknown = (
-          node.constructorParams[2] as PlanServiceNode | ConstructorNoParamNode
-        ).resolve(params);
-        const resolvedValue3: unknown = (
-          node.constructorParams[3] as PlanServiceNode | ConstructorNoParamNode
-        ).resolve(params);
 
-        return resolveFour(
+        return resolveTwo(
           resolvedValue0,
           resolvedValue1,
-          resolvedValue2,
-          resolvedValue3,
-          (
-            resolvedValue0: unknown,
-            resolvedValue1: unknown,
-            resolvedValue2: unknown,
-            resolvedValue3: unknown,
-          ): TActivated =>
-            new node.binding.implementationType(
-              resolvedValue0,
-              resolvedValue1,
-              resolvedValue2,
-              resolvedValue3,
-            ),
+          (resolvedValue0: unknown, resolvedValue1: unknown): TActivated =>
+            new node.binding.implementationType(resolvedValue0, resolvedValue1),
         );
       };
     }
@@ -81,32 +63,17 @@ export function buildFourConstructorArgumentResolverOnCsp<TActivated>(
       const resolvedValue1: unknown = (
         node.constructorParams[1] as PlanServiceNode | ConstructorNoParamNode
       ).resolve(params);
-      const resolvedValue2: unknown = (
-        node.constructorParams[2] as PlanServiceNode | ConstructorNoParamNode
-      ).resolve(params);
-      const resolvedValue3: unknown = (
-        node.constructorParams[3] as PlanServiceNode | ConstructorNoParamNode
-      ).resolve(params);
 
-      return resolveFour(
+      return resolveTwo(
         resolvedValue0,
         resolvedValue1,
-        resolvedValue2,
-        resolvedValue3,
         (
           resolvedValue0: unknown,
           resolvedValue1: unknown,
-          resolvedValue2: unknown,
-          resolvedValue3: unknown,
         ): Resolved<TActivated> =>
           resolveActivations(
             params,
-            new node.binding.implementationType(
-              resolvedValue0,
-              resolvedValue1,
-              resolvedValue2,
-              resolvedValue3,
-            ),
+            new node.binding.implementationType(resolvedValue0, resolvedValue1),
           ),
       );
     };
@@ -139,31 +106,19 @@ export function buildFourConstructorArgumentResolverOnCsp<TActivated>(
       const resolvedValue1: unknown = (
         node.constructorParams[1] as PlanServiceNode | ConstructorNoParamNode
       ).resolve(params);
-      const resolvedValue2: unknown = (
-        node.constructorParams[2] as PlanServiceNode | ConstructorNoParamNode
-      ).resolve(params);
-      const resolvedValue3: unknown = (
-        node.constructorParams[3] as PlanServiceNode | ConstructorNoParamNode
-      ).resolve(params);
 
-      return resolveFour(
+      return resolveTwo(
         resolvedValue0,
         resolvedValue1,
-        resolvedValue2,
-        resolvedValue3,
         (
           resolvedValue0: unknown,
           resolvedValue1: unknown,
-          resolvedValue2: unknown,
-          resolvedValue3: unknown,
         ): Resolved<TActivated> =>
           finalizeInstance(
             params,
             new node.binding.implementationType(
               resolvedValue0,
               resolvedValue1,
-              resolvedValue2,
-              resolvedValue3,
             ) as SyncResolved<TActivated> & Record<string | symbol, unknown>,
           ),
       );
@@ -196,31 +151,19 @@ export function buildFourConstructorArgumentResolverOnCsp<TActivated>(
     const resolvedValue1: unknown = (
       node.constructorParams[1] as PlanServiceNode | ConstructorNoParamNode
     ).resolve(params);
-    const resolvedValue2: unknown = (
-      node.constructorParams[2] as PlanServiceNode | ConstructorNoParamNode
-    ).resolve(params);
-    const resolvedValue3: unknown = (
-      node.constructorParams[3] as PlanServiceNode | ConstructorNoParamNode
-    ).resolve(params);
 
-    return resolveFour(
+    return resolveTwo(
       resolvedValue0,
       resolvedValue1,
-      resolvedValue2,
-      resolvedValue3,
       (
         resolvedValue0: unknown,
         resolvedValue1: unknown,
-        resolvedValue2: unknown,
-        resolvedValue3: unknown,
       ): Resolved<TActivated> =>
         finalizeInstance(
           params,
           new node.binding.implementationType(
             resolvedValue0,
             resolvedValue1,
-            resolvedValue2,
-            resolvedValue3,
           ) as SyncResolved<TActivated> & Record<string | symbol, unknown>,
         ),
     );
